+++ terraform/api-management.tf
@@ -1,5 +1,10 @@
+locals {
+  api_management_resource_count = var.azurerm_api_management_services == 1 ? 1 : 0
+  api_management_apis_resource_count = var.azurerm_api_management_services == 1 && var.azurerm_api_management_apis == 1 ? 1 : 0
+}
+
 resource "azurerm_api_management" "j1dev" {
-  count               = "${var.azurerm_api_management_services == 1 ? 1 : 0}"
+  count               = local.api_management_resource_count
   name                = "j1dev"
   location            = azurerm_resource_group.j1dev.location
   resource_group_name = azurerm_resource_group.j1dev.name
@@ -9,8 +14,25 @@ resource "azurerm_api_management" "j1dev" {
   sku_name = "Developer_1"
 }
 
+resource "azurerm_monitor_diagnostic_setting" "j1dev_api_mgmt_diag_set" {
+  count              = local.api_management_resource_count
+  name               = "j1dev_api_mgmt_diag_set"
+  target_resource_id = azurerm_api_management.j1dev[0].id
+  storage_account_id = azurerm_storage_account.j1dev.id
+
+  log {
+    category = "GatewayLogs"
+    enabled  = true
+
+    retention_policy {
+      enabled = true
+      days    = 1
+    }
+  }
+}
+
 resource "azurerm_api_management_api" "example" {
-  count               = "${var.azurerm_api_management_services == 1 ? (var.azurerm_api_management_apis == 1 ? 1 : 0) : 0}"
+  count               = local.api_management_resource_count
   name                = "j1dev-api"
   resource_group_name = azurerm_resource_group.j1dev.name
   api_management_name = azurerm_api_management.j1dev[0].name